third_party/blink/web_tests/fast/dom/Window/open-invalid-arguments.html

<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<script>
description("Checks that passing invalid strings to window.open() is properly handled.");

if (window.testRunner)
    testRunner.setPopupBlockingEnabled(false);

var nonConvertibleToString = { toString: function() { throw "Exception in toString()"; } };
var newWindow = null;
shouldThrow("newWindow = window.open(nonConvertibleToString)", "'Exception in toString()'");
shouldBeNull("newWindow");
shouldThrow("newWindow = window.open('about:blank', nonConvertibleToString)", "'Exception in toString()'");
shouldBeNull("newWindow");
shouldThrow("newWindow = window.open('about:blank', 'frameName', nonConvertibleToString)", "'Exception in toString()'");
shouldBeNull("newWindow");
</script>
</body>
</html>
