third_party/WebKit/LayoutTests/fast/dom/Window/open-invalid-arguments-expected.txt

Checks that passing invalid strings to window.open() is properly handled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS newWindow = window.open(nonConvertibleToString) threw exception Exception in toString().
PASS newWindow is null
PASS newWindow = window.open('about:blank', nonConvertibleToString) threw exception Exception in toString().
PASS newWindow is null
PASS newWindow = window.open('about:blank', 'frameName', nonConvertibleToString) threw exception Exception in toString().
PASS newWindow is null
PASS successfullyParsed is true

TEST COMPLETE
